Restrict delete and patch of abbreviation to user who created it

+++ src/main/java/com/abbrevio/abbrevio/service/impl/AbbreviationServiceImpl.java
@@ -2,10 +2,15 @@ package com.abbrevio.abbrevio.service.impl;
 
 import com.abbrevio.abbrevio.dto.AbbreviationDTO;
 import com.abbrevio.abbrevio.entity.Abbreviation;
+import com.abbrevio.abbrevio.entity.User;
 import com.abbrevio.abbrevio.exception.CustomNotFoundException;
 import com.abbrevio.abbrevio.repository.AbbreviationRepository;
+import com.abbrevio.abbrevio.repository.UserRepository;
 import com.abbrevio.abbrevio.service.AbbreviationService;
+import com.abbrevio.abbrevio.service.AuthService;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -15,17 +20,25 @@ import java.util.stream.Collectors;
 public class AbbreviationServiceImpl implements AbbreviationService {
 
     private final AbbreviationRepository abbreviationRepository;
+    private final UserRepository userRepository;
     private final ModelMapper modelMapper;
 
-    public AbbreviationServiceImpl(AbbreviationRepository abbreviationRepository, ModelMapper modelMapper) {
+    public AbbreviationServiceImpl(AbbreviationRepository abbreviationRepository, ModelMapper modelMapper, UserRepository userRepository) {
         this.abbreviationRepository = abbreviationRepository;
         this.modelMapper = modelMapper;
+        this.userRepository = userRepository;
     }
 
     @Override
     public AbbreviationDTO createAbbreviation(AbbreviationDTO abbreviationDTO) {
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = userRepository.findByUsername(authentication.getName())
+                .orElseThrow(() -> new CustomNotFoundException(User.class, "username", authentication.getName()));
+
         Abbreviation abbreviation = new Abbreviation();
         abbreviation.setName(abbreviationDTO.getName());
+        abbreviation.setUser(user);
         abbreviationRepository.save(abbreviation);
 
         return modelMapper.map(abbreviation, AbbreviationDTO.class);
@@ -45,9 +58,18 @@ public class AbbreviationServiceImpl implements AbbreviationService {
     }
 
     @Override
-    public AbbreviationDTO updateAbbreviation(Long id, AbbreviationDTO abbreviationDTO) {
+    public AbbreviationDTO updateAbbreviation(Long id, AbbreviationDTO abbreviationDTO) throws Exception {
         Abbreviation abbreviation = abbreviationRepository.findById(id)
                 .orElseThrow(() -> new CustomNotFoundException(Abbreviation.class, "id", id));
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = userRepository.findByUsername(authentication.getName())
+                .orElseThrow(() -> new CustomNotFoundException(User.class, "username", authentication.getName()));
+
+        if (!abbreviation.getUser().getId().equals(user.getId())) {
+            throw new Exception(String.format("you cannot edit abbreviation with id %s, because you have not created it", id));
+        }
+
 
         if (abbreviationDTO.getName() != null) {
             abbreviation.setName(abbreviationDTO.getName());
@@ -57,9 +79,23 @@ public class AbbreviationServiceImpl implements AbbreviationService {
     }
 
     @Override
-    public void deleteAbbreviation(Long id) {
+    public void deleteAbbreviation(Long id) throws Exception {
         Abbreviation abbreviation = abbreviationRepository.findById(id)
                 .orElseThrow(() -> new CustomNotFoundException(Abbreviation.class, "id", id));
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = userRepository.findByUsername(authentication.getName())
+                .orElseThrow(() -> new CustomNotFoundException(User.class, "username", authentication.getName()));
+
+        if (!abbreviation.getUser().getId().equals(user.getId())) {
+            throw new Exception(String.format("you cannot delete abbreviation with id %s, because you have not created it", id));
+        }
         abbreviationRepository.delete(abbreviation);
+    }
+
+    @Override
+    public List<AbbreviationDTO> getAllContainingName(String name) {
+        List<Abbreviation> abbreviations = abbreviationRepository.findByNameContaining(name);
+
+        return abbreviations.stream().map((abbreviation -> modelMapper.map(abbreviation, AbbreviationDTO.class))).collect(Collectors.toList());
     }
 }
